More logging and docs
- Logging has been added to all files except NOAA and TS. Will continue
to add more for next update
- Changed a lot of docs to match standard.

+++ Python/lipd/pkg_resources/doi/doi_main.py
@@ -4,26 +4,25 @@ from ..helpers.directory import *
 from ..helpers.zips import *
 from ..helpers.jsons import *
 from ..helpers.loggers import *
-from ..helpers.log_init import create_logger
 
-
-logger_doi_main = create_logger(__name__)
+logger_doi_main = create_logger("doi_main")
 
 
 def doi():
     """
     Main function that controls the script. Take in directory containing the .lpd file(s). Loop for each file.
-    :return: None
+    :return None:
     """
-    logger_doi_main.info("enter doi_main()")
+    logger_doi_main.info("enter doi_main")
     # Find all .lpd files in current directory
     # dir: ? -> dir_root
     dir_root = os.getcwd()
     f_list = list_files('.lpd')
     print("Found {0} {1} file(s)".format(str(len(f_list)), 'LiPD'))
-
+    force = prompt_force()
     for name_ext in f_list:
         print('processing: {}'.format(name_ext))
+        logger_doi_main.info("processing: {}".format(name_ext))
 
         # .lpd name w/o extension
         name = os.path.splitext(name_ext)[0]
@@ -32,12 +31,13 @@ def doi():
         dir_tmp = create_tmp_dir()
         unzip(name_ext, dir_tmp)
 
-        # Unbag and check resolved flag. Don't run if flag exists
-        if resolved_flag(open_bag(os.path.join(dir_tmp, name))):
-            # print("DOI previously resolved. Next file...")
-            shutil.rmtree(dir_tmp)
+        # Force DOI update?
+        if not force:
+            # Unbag and check resolved flag. Don't run if flag exists.
+            if resolved_flag(open_bag(os.path.join(dir_tmp, name))):
+                shutil.rmtree(dir_tmp)
 
-        # Process file if flag does not exist
+        # Process file if flag does not exist or force.
         else:
             # dir: dir_root -> dir_tmp
             process_lpd(name, dir_tmp)
@@ -48,22 +48,21 @@ def doi():
             os.rename(name_ext + '.zip', name_ext)
             # Cleanup and remove tmp directory
             shutil.rmtree(dir_tmp)
-    logger_doi_main.info("exit doi_main()")
+    logger_doi_main.info("exit doi_main")
     return
 
 
 def process_lpd(name, dir_tmp):
     """
     Opens up json file, invokes doi_resolver, closes file, updates changelog, cleans directory, and makes new bag.
-    :param name: (str) Name of current .lpd file
-    :param dir_tmp: (str) Path to tmp directory
-    :return: none
+    :param str name: Name of current .lpd file
+    :param str dir_tmp: Path to tmp directory
+    :return none:
     """
-
+    logger_doi_main.info("enter process_lpd")
     dir_root = os.getcwd()
     dir_bag = os.path.join(dir_tmp, name)
     dir_data = os.path.join(dir_bag, 'data')
-    valid = True
 
     # Navigate down to jLD file
     # dir : dir_root -> dir_data
@@ -72,20 +71,50 @@ def process_lpd(name, dir_tmp):
     # Open jld file and read in the contents. Execute DOI Resolver.
     jld_data = read_json_from_file(os.path.join(dir_data, name + '.jsonld'))
 
-    if valid:
-        # Overwrite data with new data
-        jld_data = DOIResolver(dir_root, name, jld_data).main()
-        # Open the jld file and overwrite the contents with the new data.
-        write_json_to_file(os.path.join(dir_data, name + '.jsonld'), jld_data)
+    # Overwrite data with new data
+    jld_data = DOIResolver(dir_root, name, jld_data).main()
+    # Open the jld file and overwrite the contents with the new data.
+    write_json_to_file(os.path.join(dir_data, name + '.jsonld'), jld_data)
 
-        # Open changelog. timestamp it. Prompt user for short description of changes. Close and save
-        update_changelog()
+    # Open changelog. timestamp it. Prompt user for short description of changes. Close and save
+    update_changelog()
 
     # Delete old bag files, and move files to bag root for re-bagging
     # dir : dir_data -> dir_bag
     dir_cleanup(dir_bag, dir_data)
     finish_bag(dir_bag)
 
+    logger_doi_main.info("exit process_lpd")
     return
 
 
+def prompt_force():
+    """
+    Ask the user if they want to force update files that were previously resolved
+    :return bool: response
+    """
+    logger_doi_main.info("enter prompt_force")
+    force = False
+    count = 0
+    print("Do you want to force updates for previously resolved files? (y/n)")
+    while True:
+        f = input("> ")
+        try:
+            if count == 2:
+                force = False
+                break
+            elif f.lower() in ('y', 'yes'):
+                force = True
+                break
+            elif f.lower() in ('n', 'no'):
+                force = False
+                break
+            else:
+                print("invalid response")
+        except AttributeError as e:
+            print("invalid response")
+            logger_doi_main.warn("invalid response: {}, {}".format(f, e))
+        count += 1
+    logger_doi_main("force update: {}".format(force))
+    logger_doi_main.info("exit prompt_force")
+    return force
